+++ extensions/gitlab/src/gitlabapi.ts
@@ -327,6 +327,16 @@ export class User {
   public web_url = "";
 }
 
+export class TemplateSummary {
+  public id = "";
+  public name = "";
+}
+
+export class TemplateDetail {
+  public name = "";
+  public content = "";
+}
+
 export interface Status {
   emoji: string;
   message: string;
@@ -585,6 +595,30 @@ export class GitLab {
       }));
     });
     return items;
+  }
+
+  async getProjectMergeRequestTemplates(projectId: number): Promise<TemplateSummary[]> {
+    const items: TemplateSummary[] = await this.fetch(`projects/${projectId}/templates/merge_requests`).then(
+      (templates) => {
+        return templates.map((template: any) => ({
+          id: template.key,
+          name: template.name,
+        }));
+      }
+    );
+    return items;
+  }
+
+  async getProjectMergeRequestTemplate(projectId: number, templateName: string): Promise<TemplateDetail> {
+    const item: TemplateDetail = await this.fetch(
+      `projects/${projectId}/templates/merge_requests/${templateName}`
+    ).then((template) => {
+      return {
+        name: template.name,
+        content: template.content,
+      };
+    });
+    return item;
   }
 
   async getGroupMilestones(group: Group): Promise<Milestone[]> {